+++ aks_cluster/outputs.tf
@@ -5,3 +5,7 @@ output "id" {
 output "kube_admin_config" {
   value = azurerm_kubernetes_cluster.cluster.kube_admin_config
 }
+
+output "kube_config" {
+  value = azurerm_kubernetes_cluster.cluster.kube_config
+}