+++ aks_cluster/variables.tf
@@ -15,9 +15,19 @@ variable "node_count" {
   description = "The default node pool instance count"
 }
 
+variable "enable_aad_auth" {
+  default     = false
+  description = "Enable Kubernetes API Azure AD authentication"
+}
+
 variable "enable_auto_scaling" {
   default     = true
   description = "Enable autoscaling on the default node pool"
+}
+
+variable "enable_rbac" {
+  default     = true
+  description = "Enable RBAC on the Kubernetes API"
 }
 
 variable "enable_node_public_ip" {