+++ aks_cluster/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## aks_cluster-v0.7.0
+
+### Added
+- AAD integration is now optional and defaults to false, new variable `enable_aad_auth`
+- Additional outputs for `kube_config`
+- Created examples
+- Updated `azurerm` provider to `v2.5.0`
+
+### Removed
+- A service principal password is no longer required. The module now uses `identity {type="SystemAssigned}` by default.
+
+### Fixed
+- conditional logic on optional oms_agent now works
+
 ## aks_cluster-v0.6.0
 
 ### Fixed